Begin work on nginxconf.sh -- to add web confs for post-install installations of nginx

+++ scripts/install/autodl.sh
@@ -61,11 +61,13 @@ gui-server-password = ${IRSSI_PASS}
 ADC
       chown -R $u: /home/${u}/.autodl/
       chown -R $u: /home/${u}/.irssi/
-sed -i '/?>/d' /srv/rutorrent/conf/users/${u}/config.php
-echo "\$autodlPort = \"$IRSSI_PORT\";" >> /srv/rutorrent/conf/users/${u}/config.php
-echo "\$autodlPassword = $IRSSI_PASS;" >> /srv/rutorrent/conf/users/${u}/config.php
-echo "?>" >> /srv/rutorrent/conf/users/${u}/config.php
-
+  if [[ -f /install/.nginx.lock ]]; then
+    sed -i '/?>/d' /srv/rutorrent/conf/users/${u}/config.php
+    echo "\$autodlPort = \"$IRSSI_PORT\";" >> /srv/rutorrent/conf/users/${u}/config.php
+    echo "\$autodlPassword = $IRSSI_PASS;" >> /srv/rutorrent/conf/users/${u}/config.php
+    echo "?>" >> /srv/rutorrent/conf/users/${u}/config.php
+  fi
+  
 cat >"/etc/systemd/system/irssi@.service"<<ADC
 [Unit]
 Description=AutoDL IRSSI
